Adds await on a tuple of futures.
Currently the Process::await implementation on list<Future>, by the nature of std::list, expects equally typed futures. This new override implements await for a tuple of futures, hence allows awaiting differently typed futures in a single call.

There also is a new override that allows await on a single Future, a convenience approach for allowing a Process based await on a single Future without forcing the user to render a list or tuple out of that.

A C++11 and a boost-based implementation have been added.

This patch also includes tests on those new overrides.

Review: https://reviews.apache.org/r/18311