Fixed orphaned container handling in the ExternalContainerizer recover implementation.
An orphaned container is known to the ECP but not to the EC, thus not recoverable but pending. This patch enforces a call to destroy for any orphan that has been identified as such during the recovery phase.

NOTE: Such destroy is wrapped by a call to "wait" to make sure the ExternalContainerizer gets to know when a container was destroyed successfully.

Review: https://reviews.apache.org/r/21424