Added test-containerizer cpp implementation.
The test-containerizer is posing as an external containerizer program (ECP) by using the MesosContainerizerProcess (MC) to handle the actual containerizing.
The MC is wrapped by a protobuf process based communication interface (thunk). Both processes are run in form of a daemon.
The daemon is controlled by invocations of this test-containerizer as a trigger executable.

Both processes are spawned implicitly whenever "launch" is invoked on a MESOS_WORK_DIRECTORY that does not have a daemon running yet.
The daemon terminates itself whenever there are no more active containers according to the MesosContainerizerProcess.

The resulting communication scheme is as follows:
EC <> trigger <> thunk <> MesosContainerizerProcess

This patch re-enables the ExternalContainerizer.Launch test.

Review: https://reviews.apache.org/r/21695