Re-enabled health check tests
Currently the health check tests are disabled as it was flaky. Part of the reason is that all assertions was EXPECTs instead of ASSERTs so it will continue to execute.
Another issue was it's currently discovering the path to launch health check executable from the command executor main argv[0] path.
However, the correct path is to launch the generated script that does library path resolution, which is the parent of .libs.

Review: https://reviews.apache.org/r/23443